+++ index.js
@@ -1,0 +1,8 @@
+
+import * as myclass from "./modules/class";
+import eventslistner from "./modules/function";
+import currenttimeanddate from "./modules/time";
+
+myclass();
+eventslistner();
+currenttimeanddate()
